hotfix hot bugfix for matplotlib streamlit issues
