+++ requirements.txt
@@ -1,4 +1,5 @@
 altair==4.1.0
+altgraph==0.17
 appnope==0.1.2
 argon2-cffi==20.1.0
 astor==0.8.1
@@ -8,6 +9,8 @@ backcall==0.2.0
 base58==2.1.0
 bleach==3.3.0
 blinker==1.4
+boto3==1.17.27
+botocore==1.20.27
 cachetools==4.2.1
 certifi==2020.12.5
 cffi==1.14.5
@@ -19,10 +22,12 @@ cycler==0.10.0
 decorator==4.4.2
 defusedxml==0.7.1
 entrypoints==0.3
+enum-compat==0.0.3
 et-xmlfile==1.0.1
 Fiona==1.8.18
+geojson==2.5.0
 geopandas==0.9.0
-gitdb==4.0.6
+gitdb==4.0.5
 GitPython==3.1.14
 greenlet==1.0.0
 idna==2.10
@@ -31,8 +36,10 @@ ipykernel==5.5.0
 ipython==7.21.0
 ipython-genutils==0.2.0
 ipywidgets==7.6.3
+jdcal==1.4.1
 jedi==0.18.0
 Jinja2==2.11.3
+jmespath==0.10.0
 joblib==1.0.1
 jsonschema==3.2.0
 jupyter-client==6.1.12
@@ -40,6 +47,7 @@ jupyter-core==4.7.1
 jupyterlab-pygments==0.1.2
 jupyterlab-widgets==1.0.0
 kiwisolver==1.3.1
+macholib==1.14
 MarkupSafe==1.1.1
 matplotlib==3.3.4
 mistune==0.8.4
@@ -48,18 +56,19 @@ nbclient==0.5.3
 nbconvert==6.0.7
 nbformat==5.1.2
 nest-asyncio==1.5.1
-notebook==6.3.0
+notebook==6.2.0
 numpy==1.20.1
 openpyxl==3.0.7
 packaging==20.9
 pandas==1.2.3
 pandocfilters==1.4.3
 parso==0.8.1
+pathtools==0.1.2
 pexpect==4.8.0
 pickleshare==0.7.5
 Pillow==8.1.2
 prometheus-client==0.9.0
-prompt-toolkit==3.0.18
+prompt-toolkit==3.0.17
 protobuf==3.15.6
 psycopg2-binary==2.8.6
 ptyprocess==0.7.0
@@ -67,24 +76,27 @@ pyarrow==3.0.0
 pycparser==2.20
 pydeck==0.6.1
 Pygments==2.8.1
+pyinstaller-hooks-contrib==2021.1
 pyparsing==2.4.7
 pyproj==3.0.1
 pyrsistent==0.17.3
 python-dateutil==2.8.1
 pytz==2021.1
+PyYAML==5.4.1
 pyzmq==22.0.3
 requests==2.25.1
+s3transfer==0.3.4
 scikit-learn==0.24.1
-scipy==1.6.2
+scipy==1.6.1
 seaborn==0.11.1
 Send2Trash==1.5.0
 Shapely==1.7.1
 six==1.15.0
 sklearn==0.0
 smmap==3.0.5
-SQLAlchemy==1.4.3
+SQLAlchemy==1.4.0
 streamlit==0.79.0
-terminado==0.9.3
+terminado==0.9.2
 testpath==0.4.4
 threadpoolctl==2.1.0
 toml==0.10.2
@@ -95,8 +107,10 @@ typing-extensions==3.7.4.3
 tzlocal==2.1
 urllib3==1.26.4
 validators==0.18.2
+watchdog==2.0.2
 wcwidth==0.2.5
 webencodings==0.5.1
 widgetsnbextension==3.5.1
+xlrd==2.0.1
 XlsxWriter==1.3.7
 zipp==3.4.1
